Regenerate session when user returns from CAS Auth
The CAS single-sign-out protocol POSTs an XML document describing the user being logged out. The only identifier in the document is the CAS Service Ticket issued to the user during their initial login/validation.  To log the user out of the app during SSO, we need to find the user's session from the session store and delete it. MemoryStore provides a `.all` method, which returns all sessions in the store; we can then iterate those and look for the ticket. `.all` is an optional method, and `connect-redis` does not provide it. Thus, we need to:

- Implement a custom `genid` function to generate a session ID containing the service ticket (`cas_session:::{SERVICE_TICKET}`) if there is a ticket param in the requst; otherwise, use the standard `uid(24)`.
- When the user returns from CAS and is validated, call `req.session.regenerate` to discard the old UID-based session and start a new session with a ST-based ID.